refactor(rccl): extract shared CI utilities into rccl_ci_utils.py
Move 6 identical functions from test_pytorch_c10d.py and
test_jax_collective.py into a shared module: find_rccl_library,
verify_rccl_override, parse_junit_xml, write_github_summary,
set_github_output, and send_email_report (parameterized with
subject_prefix).

Also removes temporary debug/workaround code from the JAX script
(diagnose_shared_libs, populate_rocm_lib_dir, create_soname_symlinks,
find_lib_dirs) that is no longer needed now that rocm-sdk-core and
rocm-sdk-libraries pip packages provide the ROCm runtime libraries.
